#67 Move node examples to node folder
