+++ package.json
@@ -8,7 +8,7 @@
     "lint": "eslint --ext .ts src/** && prettier src/** --check",
     "lint-fix": "eslint --ext .ts --fix src/** && prettier src/** --write",
     "node-example": "nodemon",
-    "node-example-client": "ts-node --project tsconfig.json ./src/examples/simple-node-client.ts",
+    "node-example-client": "ts-node --project tsconfig.json ./src/node/examples/simple-node-client.ts",
     "build": "rm -rf dist && tsc -p tsconfig.json",
     "test": "jest",
     "tdd": "jest --watchAll",
